Added combo boxes into table views

+++ src/main/java/gui/controls/MainWindowController.java
@@ -1,7 +1,10 @@
 package gui.controls;
 
-import gui.controls.TableCellComponent.DateEditingCell;
-import gui.controls.TableCellComponent.EditingCell;
+import gui.controls.tableCellComponent.ComboBoxCell;
+import gui.controls.tableCellComponent.DateEditingCell;
+import gui.controls.tableCellComponent.EditingCell;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.ObservableList;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
@@ -16,10 +19,7 @@ import java.sql.Date;
 import java.sql.Time;
 import java.sql.Timestamp;
 import java.time.ZoneId;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 
 
 public class MainWindowController {
@@ -66,42 +66,51 @@ public class MainWindowController {
             changedList.clear();
         });
         deleteBtn.setOnAction(e -> {
-            Tab selectedTab = mainPane.getSelectionModel().getSelectedItem();
-            TableView table = (TableView) selectedTab.getContent().lookup("#" + selectedTab.getId());
-            Service service = serviceMap.get(table);
-            if (!table.getSelectionModel().isEmpty()) {
-                service.delete(table.getSelectionModel().getSelectedItem());
-                table.setItems(service.findAll(table.getId()));
+            Alert alert = new Alert(Alert.AlertType.WARNING, "Are you sure?!");
+            alert.setHeaderText("Warning");
+            ButtonType buttonTypeDelete = new ButtonType("Delete");
+            ButtonType buttonTypeCancel = new ButtonType("Cancel");
+            alert.getButtonTypes().setAll(buttonTypeDelete, buttonTypeCancel);
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.get() == buttonTypeDelete) {
+                Tab selectedTab = mainPane.getSelectionModel().getSelectedItem();
+                TableView table = (TableView) selectedTab.getContent().lookup("#" + selectedTab.getId());
+                Service service = serviceMap.get(table);
+                if (!table.getSelectionModel().isEmpty()) {
+                    service.delete(table.getSelectionModel().getSelectedItem());
+                    table.setItems(service.findAll(table.getId()));
+                }
+                else
+                {
+                    Alert alert_ = new Alert(Alert.AlertType.ERROR, "An exception occurred!");
+                    alert_.setHeaderText("Error");
+                    alert_.showAndWait();
+                }
             }
         });
         addBtn.setOnAction(e -> {
             Tab selectedTab = mainPane.getSelectionModel().getSelectedItem();
             TableView table = (TableView) selectedTab.getContent().lookup("#" + selectedTab.getId());
-            if(table == airportTable)
-            {
+            if (table == airportTable) {
                 airportTable.getItems().add(new Airport());
             }
-            if(table == boardTable)
-            {
+            if (table == boardTable) {
                 boardTable.getItems().add(new Board());
             }
-            if(table == clientTable)
-            {
+            if (table == clientTable) {
                 clientTable.getItems().add(new Client());
             }
-            if(table == flightTable)
-            {
+            if (table == flightTable) {
                 flightTable.getItems().add(new Flight());
             }
-            if(table == ticketTable)
-            {
+            if (table == ticketTable) {
                 ticketTable.getItems().add(new Ticket());
             }
-            if(table == flightInfoTable)
-            {
+            if (table == flightInfoTable) {
                 flightInfoTable.getItems().add(new FlightInfo());
             }
         });
+        //add alert for changing tabs!
     }
 
     private void RefreshAll() {
@@ -120,17 +129,20 @@ public class MainWindowController {
         setFlightInfoTable();
         setTicketTable();
     }
+
     private void setAirportTable() {
         TableColumn column = createAirportCityColumn();
         TableColumn column1 = createAirportCodeColumn();
         airportTable.getColumns().addAll(column, column1);
     }
+
     private void setBoardTable() {
         TableColumn column = createBoardLastRepairColumn();
         TableColumn column1 = createBoardJetTypeColumn();
         TableColumn column2 = createBoardFreeSeatColumn();
         boardTable.getColumns().addAll(column, column1, column2);
     }
+
     private void setFlightTable() {
         TableColumn column = createFlightDateColumn();
         TableColumn column1 = createFlightDepartureColumn();
@@ -150,10 +162,12 @@ public class MainWindowController {
 
         return airportCodeColumn;
     }
-    private TableColumn createAirportCityColumn() {TableColumn city = new TableColumn("City");
+
+    private TableColumn createAirportCityColumn() {
+        TableColumn city = new TableColumn("City");
         city.setMinWidth(100);
         city.setCellValueFactory(new PropertyValueFactory<Airport, String>("city"));
-        city.setCellFactory(p -> new EditingCell<Airport,String>());
+        city.setCellFactory(p -> new EditingCell<Airport, String>());
         city.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Airport, String>>) t -> {
             t.getRowValue().setCity(t.getNewValue());
             changedSet.add(t.getRowValue());
@@ -176,23 +190,26 @@ public class MainWindowController {
 
         return boardLastRepairColumn;
     }
+
     private TableColumn createBoardFreeSeatColumn() {
 
         TableColumn boardFreeSeatColumn = new TableColumn("Free seat");
         boardFreeSeatColumn.setMinWidth(100);
-        boardFreeSeatColumn.setCellValueFactory(new PropertyValueFactory<Board, Integer>("freeSeat"));
-        boardFreeSeatColumn.setCellFactory(p -> new EditingCell<Board,Integer>());
-        boardFreeSeatColumn.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Board, Integer>>) t -> {
-            t.getRowValue().setFreeSeat(t.getNewValue());
+        boardFreeSeatColumn.setCellValueFactory(new PropertyValueFactory<Board, String>("freeSeat"));
+        boardFreeSeatColumn.setCellFactory(p -> new EditingCell<Board, String>());
+        boardFreeSeatColumn.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Board, String>>) t -> {
+            //check positive int
+            t.getRowValue().setFreeSeat(Integer.parseInt(t.getNewValue()));
             changedSet.add(t.getRowValue());
         });
         return boardFreeSeatColumn;
     }
+
     private TableColumn createBoardJetTypeColumn() {
         TableColumn boardJetTypeColumn = new TableColumn("Jet type");
         boardJetTypeColumn.setMinWidth(100);
         boardJetTypeColumn.setCellValueFactory(new PropertyValueFactory<Board, String>("jetType"));
-        boardJetTypeColumn.setCellFactory(p -> new EditingCell<Board,String>());
+        boardJetTypeColumn.setCellFactory(p -> new EditingCell<Board, String>());
         boardJetTypeColumn.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Board, String>>) t -> {
             t.getRowValue().setJetType(t.getNewValue());
             changedSet.add(t.getRowValue());
@@ -211,18 +228,20 @@ public class MainWindowController {
         });
         return flightDestinationColumn;
     }
-    private TableColumn createFlightDateColumn(){
+
+    private TableColumn createFlightDateColumn() {
         TableColumn flightDateColumn = new TableColumn("Date");
         flightDateColumn.setMinWidth(100);
         flightDateColumn.setCellValueFactory(new PropertyValueFactory<Flight, Date>("date"));
         flightDateColumn.setCellFactory(p -> new DateEditingCell<Flight, java.util.Date>());
-        flightDateColumn.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Flight, Date>>) t -> {
-            t.getRowValue().setDate(t.getNewValue());
+        flightDateColumn.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Flight, java.util.Date>>) t -> {
+            t.getRowValue().setDate(Date.valueOf(t.getNewValue().toInstant().atZone(ZoneId.systemDefault()).toLocalDate()));
             changedSet.add(t.getRowValue());
         });
         return flightDateColumn;
     }
-    private TableColumn createFlightDepartureColumn(){
+
+    private TableColumn createFlightDepartureColumn() {
         TableColumn flightDepartureColumn = new TableColumn("Departure");
         flightDepartureColumn.setMinWidth(100);
         flightDepartureColumn.setCellValueFactory(new PropertyValueFactory<Flight, Airport>("departure"));
@@ -233,11 +252,12 @@ public class MainWindowController {
         });
         return flightDepartureColumn;
     }
+
     private TableColumn createFlightDurationColumn() {
         TableColumn flightDepartureColumn = new TableColumn("Duration");
         flightDepartureColumn.setMinWidth(100);
-        flightDepartureColumn.setCellValueFactory(new PropertyValueFactory<Flight, String >("duration"));
-        flightDepartureColumn.setCellFactory(p -> new EditingCell<Flight,String>());
+        flightDepartureColumn.setCellValueFactory(new PropertyValueFactory<Flight, String>("duration"));
+        flightDepartureColumn.setCellFactory(p -> new EditingCell<Flight, String>());
         flightDepartureColumn.setOnEditCommit((EventHandler<TableColumn.CellEditEvent<Flight, String>>) t -> {
             t.getRowValue().setDuration(Time.valueOf(t.getNewValue()));
             changedSet.add(t.getRowValue());
